Kaggle's Titanic: Machine Learning for Disaster solution
The program reads in the training and test data. Empty cells are
replaced in the age, embarked and fare columns for both data sets. An
analysis is carried out to see how features of the training data
correlate with survival. The features for each person include: age. sex.
social class, number of parents/children aboard, number of
siblings/spouses aboard, which port they embarked from  and ticket fare.
New features are generated for each possible title and family size.
The training data is used to train a random forest model using whether
or not the passenger survived as the output and the other features
mentioned previously as inputs. The model is applied to the test set and
the final solution is displayed for each test passenger.